+++ src/main/java/grid/intern/storeApp/exceptions/productExceptions/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package grid.intern.storeApp.exceptions.productExceptions;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long id) {
+    public ProductNotFoundException(Integer id) {
         super("No product found with id: " + id);
     }
 }
